Extract components into dedicated files, centralized control with context
