+++ src/lib/tusktask/hooks/context/useChatContext.ts
@@ -1,0 +1,12 @@
+import React, { useContext } from "react";
+import { ChatContext } from "../../context/ChatContext";
+
+const useChatContext = () => {
+  const context = useContext(ChatContext);
+  if (!context) {
+    throw new Error("Chat Context is out of reach!");
+  }
+  return context;
+};
+
+export default useChatContext;
