+++ src/ui/components/tusktask/prefabs/Messages/ChatBubble.tsx
@@ -1,0 +1,53 @@
+import { Message } from "@/src/lib/tusktask/context/ChatContext";
+import { motion } from "motion/react";
+import { messageVariants } from "./variants";
+
+// Chat Bubble Component
+const ChatBubble = ({ message }: { message: Message }) => {
+  const isMe = message.sender === "me";
+
+  return (
+    <motion.div
+      className={`flex ${isMe ? "justify-end" : "justify-start"} mb-4`}
+      variants={messageVariants}
+      initial="hidden"
+      animate="visible"
+      exit="exit"
+      layout
+    >
+      <div className={`max-w-[70%] ${isMe ? "order-2" : "order-1"}`}>
+        {!isMe && message.senderName && (
+          <motion.p
+            className="text-xs text-muted-foreground mb-1 px-2"
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            transition={{ delay: 0.1 }}
+          >
+            {message.senderName}
+          </motion.p>
+        )}
+        <motion.div
+          className={`px-4 py-2 rounded-2xl ${
+            isMe
+              ? "bg-primary text-primary-foreground rounded-br-md"
+              : "bg-muted text-muted-foreground rounded-bl-md"
+          }`}
+          whileHover={{ scale: 1.02 }}
+          transition={{ type: "spring", stiffness: 400, damping: 25 }}
+        >
+          <p className="text-sm">{message.text}</p>
+        </motion.div>
+        <motion.p
+          className={`text-xs text-muted-foreground mt-1 px-2 ${isMe ? "text-right" : "text-left"}`}
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          transition={{ delay: 0.2 }}
+        >
+          {message.timestamp}
+        </motion.p>
+      </div>
+    </motion.div>
+  );
+};
+
+export default ChatBubble;
